Cleanup: Move Modals from uiComponents to common/Modals
