+++ ui/src/components/common/Modals/NewModalHeader.tsx
@@ -1,8 +1,8 @@
 import { DialogTitle } from '@headlessui/react';
 import { FC } from 'react';
 import { twMerge } from 'tailwind-merge';
-import { Row } from '../../components/common/LayoutComponents';
-import { CloseIconButton } from '../CloseIconButton';
+import { CloseIconButton } from '../../../uiComponents';
+import { Row } from '../LayoutComponents';
 
 const testIds = {
   closeButton: 'ModalHeader::closeButton',
