+++ ui/src/components/routes-and-lines/common/StopsNeedingUpdateModal.tsx
@@ -3,9 +3,10 @@ import uniq from 'lodash/uniq';
 import { FC } from 'react';
 import { useTranslation } from 'react-i18next';
 import { RouteTypeOfLineEnum } from '../../../generated/graphql';
-import { Modal, ModalBody, SimpleButton } from '../../../uiComponents';
+import { SimpleButton } from '../../../uiComponents';
 import { StopsList } from '../../common/ChangeHistory';
 import { Row } from '../../common/LayoutComponents';
+import { Modal, ModalBody } from '../../common/Modals';
 import { StopMetaTypeUpdateInfo } from './useUpdateStopRegistryStopMetatype';
 
 const testIds = {
